Winners 객체 삭제
