+++ src/main/java/step5/controller/OutPutView.java
@@ -1,10 +1,10 @@
 package step5.controller;
 
 import step5.domain.Racing;
-import step5.domain.Winners;
+import step5.domain.car.Cars;
 
 public class OutPutView {
-    public void winnerIs(Winners winners) {
+    public void winnerIs(Cars winners) {
         System.out.print(winners.toStringWinner());
         System.out.println("가 최종 우승했습니다.");
